+++ source/TEST/SuffixArrayTrypticCompressed_test.C
@@ -208,7 +208,7 @@ START_SECTION(void printStatistic())
 	//only for internal use
 END_SECTION
 
-START_SECTION((void findSpec(std::vector< std::vector< std::pair< std::pair< int, int >, float > > > &candidates, const std::vector< double > &spec)))
+START_SECTION((void findSpec(std::vector< std::vector< std::pair< std::pair< int, int >, double > > > &candidates, const std::vector< double > &spec)))
 	double masse[255];
 	ResidueDB* rdb = ResidueDB::getInstance();
 		
@@ -229,7 +229,7 @@ START_SECTION((void findSpec(std::vector< std::vector< std::pair< std::pair< int
 	spec.push_back(245.2816);
 	spec.push_back(387.4392);
 	vector<double> specc(spec);
-	vector<vector<pair<pair<int, int>, float> > > res;
+	vector<vector<pair<pair<int, int>, double> > > res;
 	sa->findSpec(res, specc);
 	
 	TEST_EQUAL(res.size(),specc.size());
@@ -267,7 +267,7 @@ START_SECTION((void findSpec(std::vector< std::vector< std::pair< std::pair< int
 	vector<double> spec_new;
 	for (int i = 500; i < 5000; i += 20)
 	{
-		spec_new.push_back((float)i);
+		spec_new.push_back((double)i);
 	}
 	const vector<double> specc_new (spec_new);
 	res.clear();
@@ -295,7 +295,7 @@ START_SECTION((void findSpec(std::vector< std::vector< std::pair< std::pair< int
 		for (unsigned int j = 0;j<res.at(i).size();j++)
 		{
 			String seq = txt.substr(res.at(i).at(j).first.first,res.at(i).at(j).first.second);
-			float m = 18;
+			double m = 18;
 			for (unsigned int k = 0; k < seq.length();k++)
 			{
 				m += masse[(int)seq[k]];
@@ -386,7 +386,7 @@ START_SECTION((void findSpec(std::vector< std::vector< std::pair< std::pair< int
 		for (unsigned int j = 0;j<res.at(i).size();j++)
 		{
 			String seq = txt.substr(res.at(i).at(j).first.first,res.at(i).at(j).first.second);
-			float m = 18.0;
+			double m = 18.0;
 			for (unsigned int k = 0; k < seq.length();k++)
 			{
 				m += masse[(int)seq[k]];
